Removes local bgs refactored, local bgs proxy, error handler, misc files and related spec files. Moves find definition out of local bgs refactored directory, and ixes assoc require statements and references. Removes flipper :claims_api_local_bgs_refactor.
