+++ modules/claims_api/app/models/claims_api/power_of_attorney_request/summary/search.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'claims_api/bgs_client'
+
 module ClaimsApi
   class PowerOfAttorneyRequest
     class Summary
